hw/i386: split x86.c in multiple parts
commit b061f0598b9231f7992aff4fcdf3f336f9747d11 upstream.

Keep the basic X86MachineState definition in x86.c.  Move out functions that
are only needed by other files: x86-common.c for the pc and microvm machines,
x86-cpu.c for those used by accelerator code.

Intel-SIG: commit b061f0598b92 hw/i386: split x86.c in multiple parts.
CPU new topology backporting

Message-ID: <20240509170044.190795-11-pbonzini@redhat.com>
[ Quanxian Wang: amend commit log ]
